requirements.txt
psycopg2-binary==2.9.6
vk-api==11.9.9
